Adds dependency on python-dateutil for <3.11
Prior to Python3.11, datetime.datetime.fromisoformat() was only able to
parse a limited subset of valid ISO-8601 timestamps. Notably, it did NOT
successfully handle ',' as a decimal separator, and did NOT handle a
trailing 'Z' indicating UTC / "Zulu time".

To work around this issue, setup.cfg is updated such that for Python 3.9
and Python 3.10, the third-party module 'python-dateutil' is added as a
dependency, and 'dateutil.parser.isoparse' is used instead of
'dateutil.parser.fromisoformat'